new annotation processor
instead of modify the AST tree of the classes and add methods,
classes, etc..., I'm going to create a new class with the same name
than the original but with the suffix _.

For example: `Test` will be `Test_`

Then, the required method to convert higher kinded types to original
types, will be in the `Test_` class.

The original class should extend HigherX interface in order to make
this work.